Appointment detail flicker problem resolved (#9987)
